requirement.txt
langchain==0.0.154
PyPDF2==3.0.1
hugchat
python-dotenv==1.0.0
streamlit==1.18.1
faiss-cpu==1.7.4
streamlit-chat
streamlit-extras
altair==4 
openai
tiktoken==0.5.1
sentence_transformers
## chromadb
chromadb-client
fastapi 
urlparse
